Minibatch sensitivity analysis
Updated results using the minibatch sampling. The batch size is not tested here and it's influence remains to be seen for the reconstruction loss. The results are similar to the ones obtained when using the full-batch optimization (as the size of the batch is set to consider everything). 
TODO: Consider smaller batch sizes.